+++ docs/.vuepress/configs/sidebar/en.ts
@@ -3,22 +3,19 @@ import type { SidebarConfig } from '@vuepress/theme-default'
 export const en: SidebarConfig = {
   '/guide/': [
     {
-      text: 'Guide',
-      children: [
-        '/guide/README.md',
-        '/guide/getting-started.md',
-        '/guide/configuration.md',
-        '/guide/page.md',
-        '/guide/markdown.md',
-        '/guide/assets.md',
-        '/guide/i18n.md',
-        '/guide/deployment.md',
-        '/guide/theme.md',
-        '/guide/plugin.md',
-        '/guide/bundler.md',
-        '/guide/migration.md',
-      ],
+      text: 'Get Started',
+      children: ['/guide/README.md', '/guide/getting-started.md'],
     },
+    '/guide/configuration.md',
+    '/guide/page.md',
+    '/guide/markdown/',
+    '/guide/assets.md',
+    '/guide/i18n.md',
+    '/guide/deployment.md',
+    '/guide/theme/',
+    '/guide/plugin.md',
+    '/guide/bundler.md',
+    '/guide/migration.md',
   ],
   '/advanced/': [
     {