+++ packages/@starzkg/vuepress-plugin-markdown-enhance/docs/.vuepress/config.ts
@@ -1,0 +1,105 @@
+import type { UserConfig } from '@vuepress/cli'
+import type { DefaultThemeOptions } from '@vuepress/theme-default'
+
+const config: UserConfig<DefaultThemeOptions> = {
+  base: process.env.VuePress_BASE || '/',
+
+  head: [
+    ['link', { rel: 'icon', href: '/favicon.ico' }],
+    ['meta', { name: 'author', content: 'Mr.Hope' }],
+    [
+      'meta',
+      {
+        name: 'viewport',
+        content: 'width=device-width, initial-scale=1.0, viewport-fit=cover',
+      },
+    ],
+  ],
+
+  locales: {
+    '/': {
+      lang: 'en-US',
+      title: 'Markdown Enhance Plugin',
+
+      description: 'Markdown Enhancement for VuePress',
+    },
+  },
+
+  themeConfig: {
+    logo: '/logo.svg',
+
+    locales: {
+      '/': {
+        lang: 'en-US',
+        selectText: 'Language',
+        ariaLabel: 'Select language',
+        label: 'English',
+        editLinkText: 'Edit on Github',
+
+        navbar: [
+          { text: 'Home', link: '/' },
+          { text: 'Guide', link: '/guide/' },
+          { text: 'Config', link: '/api/' },
+        ],
+
+        sidebar: {
+          '/guide/': [
+            '/guide/README.md',
+            '/guide/sup-sub.md',
+            '/guide/align.md',
+            '/guide/footnote.md',
+            '/guide/mermaid.md',
+            '/guide/tex.md',
+            '/guide/demo/README.md',
+            '/guide/demo/normal.md',
+            '/guide/demo/vue.md',
+            '/guide/demo/react.md',
+            '/guide/presentation/README.md',
+            '/guide/presentation/demo.md',
+            '/guide/presentation/themes.md',
+          ],
+
+          '/': ['/README.md', '/guide/README.md'],
+        },
+      },
+    },
+
+    repo: 'https://github.com/mister-hope/vuepress-plugin-markdown-enhance',
+    docsDir: 'docs',
+    repoLabel: 'Github',
+    editLinks: true,
+    editLinkText: '在 GitHub 上编辑此页',
+
+    themePlugins: {
+      container: {
+        tip: false,
+        warning: false,
+        danger: false,
+        details: false,
+      },
+    },
+  },
+
+  plugins: [
+    [
+      '@starzkg/markdown-enhance',
+      {
+        enableAll: true,
+        presentation: {
+          plugins: [
+            'highlight',
+            'math',
+            'search',
+            'notes',
+            'zoom',
+            'anything',
+            'audio',
+            'chalkboard',
+          ],
+        },
+      },
+    ],
+  ],
+}
+
+export default config