+++ packages/@starzkg/vuepress-shared/src/shared/i18n/index.ts
@@ -1,0 +1,2 @@
+export type { HopeLang, ResolvedLocaleConfig } from './config'
+export * from './i18n'