+++ packages/@starzkg/vuepress-theme-star/src/shared/options.ts
@@ -53,10 +53,10 @@ export interface StarThemePluginsOptions {
 
   /**
    * Markdown 增强插件配置
-   * @see http://vuepress-theme-hope.github.io/md-enhance/zh/config/
+   * @see http://vuepress-theme-hope.github.io/markdown-enhance/zh/config/
    *
    * Markdown enhance plugin options
-   * @see http://vuepress-theme-hope.github.io/md-enhance/config/
+   * @see http://vuepress-theme-hope.github.io/markdown-enhance/config/
    */
   mdEnhance?: MarkdownEnhanceOptions | false
 }